PageUpdater: move LogEntry back to main transaction.
This is a partial revert of I48977ac0547bd7bd694397e0561dbb1ad1c9a42e.
In that change, the creation of the log entry for page creation was
moved from the atomic section in PageUpdater to a listener in
ChangeTrackingEventIngress which would be invoked in a deferred update
if the main transaction was successful.

On second though, seems undesirable for two reasons:
1) We probably want to keep strong consistency guarantees between the
revision history and the log table. The original change would have
allowed for a log entry to be lost if the request failed after the
transaction had been committed but before the deferred updates were
executed.

2) We'll want the behavior of PageUpdater to be consistent with
the DeletePage and UndeletePage commands, so we can introduce
PageDeletedEvent and PageUndeletedEvent. DeletePage provides the
log IDs as a result, so log entries have to created immediately.
Both commands expose log entries to the relevant hooks.

Bug: T377229
Change-Id: I139155786dc948f311f1d69eb93efc156516fd37