Category: Fix mistake when adjusting negative page counts
7854e75a66f83e87 added another page count check, specifically that the
total number of regular pages (total number of pages, minus the numbers
of subcategories and files) must not be negative. As with the existing
checks, this may trigger a recount of all pages in the category.

Until the recount happens, the total number of pages is adjusted to make
it consistent with this rule. However, the calculation is incorrect, in
that the number of regular pages is computed, not the total pages, which
must be at least the sum of the other page counts. Fix the calculation
to use that sum, not 0, as the minimum, instead of subtracting it from
the total number of pages (and failing to add it back afterward).

Also updated the comment, and added a couple more, to state why we are
doing this instead of just making sure that none of the three properties
is set to a negative number as had been done before.

Also removed the check for a negative total because PHP will make sure
(by converting to float) that underflow will not change the sign from
negative to positive when we subtract the numbers of subcategories and
files (which we ensure are not negative) from the total.

Bug: T373773
Change-Id: I1360abbc594d89be899a6081c84cff2f160878fa