Collect test failure logs and print them at the end of parallel runs
The introduction of parallel testing has made it harder to review
the console output of CI runs containing failures as the failure
information is distributed throughout the log output and not in a
single block at the end of the run.

Parse and collect the PHPUnit tests output from parallel runs, and
dump collected errors at the end of the run in a single block in
the case that there are failures.

Bug: T378481
Change-Id: Iafe4604c1a73e812362b234c676cd07c422417c7