TextPassDumper: refresh content address on failure
One potential failure case of the two-pass dump process is that
mediawiki doesn't guarantee long-term stability of the content
addresses, but we write the content address as part of the stub
and then look it up in the text pass some time later.

To address this failure case add a step to refresh the content
address from the database by revision/slot that runs if the
initial fetch from the database fails. This keeps the original db
query patterns in place, but adds additional queries on retry to
ensure we have the correct content address.

Bug: T377594
Change-Id: I31e39ecb3b0d14eacdc6f25d00fe3c63f6b02c1d