Remove invisible direction marks from Persian time formats
As a part of reducing rely on hidden direction marks for
the reasons explained in T375975 now let's get rid of
them in date formats inserting their own hidden RLM
characters as a part of localized date format.

There are three locale that insert RLM as a part of date
format, fa, ckb and kk-arab, but there are lots of RTL
script locales that don't do such a thing so it will be
nice to remove the hidden characters from these locales
and deal with them in markup but let's start with the
Persian one which if there won't be any issue with the
change that can't be dealt with markup and CSS, we can
remove them more safely from ckb and kk-arab that are
less impacting in comparison to Persian locale also.

These hidden characters were added in
https://github.com/wikimedia/mediawiki/commit/24b6d652c25b71e90beb5c9f5cf82b2f073520a6
mentioning a usecase which if it did matter that much it
wasn't limited only to Persian but regardless is better
to be replaced with a markup based solution now with the
advent of the web and this time finding a solution that
includes other RTL locales also.

Bug: T375975
Change-Id: Ic83116bdb35922f62dc4f82384743a95a4220f0c