SpecialBlock: Reload BlockLog components after saving a block
Add a key to the BlockLog components that changes after saving
a block, forcing these components to reload and show the new
block details and updated list of log entries.

The value of this key doesn't matter, it just has to be unique and
change when a block is saved.

Bug: T378586
Change-Id: Ie86302b6c1c4afd4c570858adbba56c115604da4