SpecialBlock [Codex]: Do not convert autoBlockExpiry to a ref
The store exposes autoBlockExpiry as a string, so use it directly
instead of via storeToRefs(). It could be removed from the store
entirely, but as the autoBlock flag is there it seems neat to
keep it alongside that (and maybe we'll want to use it in some
calculated value in the future).

Bug: T377509
Change-Id: Ideee5d84c44060fe212d4bde486be8e8ef7626e1