Parsoid DataAccess: support strip state in parser function results
When a parser function returns raw HTML in the strip state, this
patch ensures it will be properly tunnelled through Parsoid as part
of Parsoid's PFragment type.

The Parser::OT_PREPROCESS mode claims that it processes "templates
but not extension tags" but in fact it does process extension tags
and turns them into general strip state items.  Add a new parameter
to Parser::replaceVariables() which leaves extension tags untouched
(so that they can later be expanded by Parsoid).

Bug: T279094
Bug: T257606
Bug: T374661
Depends-On: I0d8e85b01435f37b8bccab3431c9807d46989ae6
Depends-On: Ia201767111872158149ad805f415fe1fbe9fa70e
Depends-On: I14f81b478ab7bfea98f825d619c492adce4f7a30
Change-Id: I52ed681d06c133574b57c97b0f5040199468597b