PagesField.vue: enforce maximum 10 pages, or 50 for multiblocks
This is a clunky solution to enforce the current maximum of 10 pages on
partial blocks. If multiblocks is enabled, it will allow up to 50 pages.

The loading state of the MultiselectLookup component sometimes is shown
when attempting to add chips past the maximum amount. This doesn't seem
to be fixable without a change in Codex (T379414). The user however is
still not allowed to add more, and the placeholder is also removed for
clarity.

Bug: T362718
Change-Id: I9049a58d687611f2a498d9341baeff3344832549