Allow other input and changes to trigger searchsuggestions to update
The searchsuggest logic was not forwarding events which might change
the state of the autocomplete. This made it so that composed charaters
and so called 'dead key' updates would not register.

While the default search is not affected, search completion in Vector,
monobook and the inputbox extension was still affected by this.

Bug: T177251
Change-Id: I889d926400f6627918583f170fe0c7d3f9704da6