ActionEntryPoint: Redirect non-local interwiki to Special:GoToInterwiki
If you manually make a request to /wiki/ and type a non-local interwiki,
such as "google:", this is currently denied with "Bad title".

While such links should be impossible to arrive at, because MediaWiki
does not make them in server-side code (messages and wikitext parsing
would resolve them and create an external link directly), this can
happen in pratice when using JavaScript mw.message() in core or an
extension, when the interface message uses such interwiki links.

Interface messages in the core software may use interwiki links, so long
as those interwiki prefixes are defined by default (e.g. "mw:"). Plus,
local overrides could use even more and those are expected to work
regardless of whether the message is used in JS or PHP.

Bug: T377640
Change-Id: I325fcaedee5b9603f14d66f434946f7d5c4ede71