installer: Virtual domain and external store installation
During installPreConfigured.php, if the database for a virtual domain
doesn't exist, create it, unless the domain is "shared" and the shared
option was not specified. Also create external store databases, and
initialise tables.

Databases with a truthy "db" parameter are counted as shared, even
though we don't really know the user's intention, since that's safest
for production.

This is for feature parity with the CREATE DATABASE queries in
addWiki.php. Databases on an external cluster with db=false always
need to be created when a new wiki is created.

Also:
* Factor out database creation to a helper which can be shared between
  main and external tables.
* Merge MysqlCreateDatabaseTask and PostgresCreateDatabaseTask since
  they are the same now.
* In DatabaseUpdater, skip updates for potentially shared virtual
  tables. This is necessary for production safety and should replace
  some extension hacks.

Bug: T352113
Change-Id: I739a39029747b023b03583d9d9c4616115401672