installer: Schedule sysop and initial content after extension tables
With the traditional installer, it's pretty harmless to create the main
page early, because we're not using extension hooks. But with
installPreConfigured.php, the extension hooks are all installed and
active, and so tables need to exist for these operations to work.

This reverts Iba74c681e5507560ade3b079157ac052adbbc304.

Depends-On: Ia3a94414f92574b3f1b267802874c4f4865972a5
Change-Id: Id5e304ae3704e4f90ec4c2d82e8d827e2629493a