SpecialBlock: Fix up display of suppressed log entries
Don't show the 'Target' column for the suppress log, and filter
out anything other than block and reblock entries by type. This
is simpler than splitting it up into two API queries, but has
the downside of not showing a full 10 entries if there are that
many and some are non-block related ones.

Bug: T377725
Change-Id: Ib725f74ed327d8ab85f38e9f764d707502c4ef0d