page: Restore call to PageUpdater::setUsePageCreationLog()
This was supposed to be restored when the page creation
log entry generation was moved back to main transaction and the
PageUpdater method was undeprecated in 2166c09f469a1059 (1085458).

Restoring this call restores the prior behavior before the regression
(of not creating extra creation log entries for redirect pages that
are created automatically as a result of an action that is already separately logged).

Bug: T379380
Change-Id: Icf0f31916dadfd54974529ef784c054789b599ce