parser: Skip most of Parser::finalizeHeadings when no header found
When the regex does not find <h1-6> tags, most of the function does not
need to initialize variables for the empty loop run.
In this case the NO_TOC output flag is no longer set.

This should improve message parsing as the short text snippet from
messages often does not provide headers

Bug: T378764
Change-Id: Ib22c40f10494f7fbe5cb27602a17a365642607e8