Add migrateSearchindex functions
This runs migrations that were previously attempted.
Postgres never executed as the renameindex failed and it never ran the
patch either.

MySql did run, but was relying on renameIndex with PRIMARY, which is a
bad idea. The function now runs it as a patch, with upgrade row in the
log, but also checks for the lack of the old index in case the users
were already on 1.44 or if they used 1.43rc

Bug: T379591
Change-Id: I7e74ccbf8cf2b105e8e6dedf812af3487c18113b
(cherry picked from commit def5b99737a296c108310881debe6ecb8bac1b2f)