SpecialBlock [Codex]: use confirmation dialog instead of checkbox
Add a reusable ConfirmationDialog, akin to OO.ui.confirm(). A local prop
to track the 'open' state is required, but otherwise you only need to
provide 'title', the content (the only slot), and what to do on
confirmation via the 'confirm' event. SpecialBlock.vue – where form
submission is handled – is responsible for opening this dialog, while
the store has a watcher to set the confirmationMessage ref.

TODOs are left to further abstract ConfirmationDialog and other
eligible components for general use in MediaWiki (T375220).

Simplify some components by putting the formSubmitted prop into the
store.

Update tests and use toStrictEqual() instead of toBe(), following the
same convention of strict equality comparisons used in JS and QUnit
tests.

Bug: T376754
Change-Id: I73d4fecb7f91e565a9029311c6d5cd9cfe7ab230