search: Move SearchUpdate.php to /includes/search directory
The class was introduced for and is exclusively used by the Search
component, and is already documented as part of the Search docgroup.

It was placed under the deferred directory because, 10+ years ago,
we used to treat MediaWiki as something that (mostly) a single team
can maintain and look after, but this has changed a lot since then.

* Mark the class as internal. I can't think of a use case outside of
  core, and indeed it has no references outside core
  (per Codesearch/Everywhere).

  Follows-up Ibd40734b9, which namespaced this class under Deferred.
  Remove both the global class alias and the Deferred alias, since
  the class is now internal and requires no back compat aliases.

Bug: T364652
Change-Id: I48937dcda4144e5e039c38a2e336f8361e2d492d