MessageCache::parseWithPostprocessing: remove $options argument
During post-merge code review, it was noted that passing
`'unwrap'=>false` in the $options array might trigger post processing
on the wrapper div, which would be a behavior change (and surprising).

None of the callers actually pass an options array yet, so remove it
as a parameter and always use the default options (which set `unwrap`
to true).  We can always add $options back later if/when needed.

Change-Id: I5c90da188298f77b549319abbd8e45fa4cdd8eb6