editpage: More consistently exclude unintended limit report entries
Some entries in the limit report are not supposed to appear in this
table, because they require parameters other than the one or two
numeric parameters supported by this code. RenderDebugInfo already has
a list of them and special handling.

However, excluding them relied on this check within the foreach loop:
if ( !$keyMsg->isDisabled() && !$valueMsg->isDisabled() ) ...
which will pass in qqx mode, or if the message key is defined on-wiki.
Instead, exclude them based on a list, like in RenderDebugInfo.

Bug: T379971
Change-Id: Ie0ccb2d7ddab0ac4ba30f27c908059f23fb387a1