ExpiryField.vue: react to changes to the store's expiry
The reactivity to expiry changes was broken in Id6cff21eee. Now that the
expiry is in the store, listen to it and react accordingly, and gaurd
against circular reactivity by ensuring the new value is different.

Change setDurationFromGiven() to handle blank values.

Add test and remove setting of modelValue (which is a relic from before
Id6cff21eee).

Fix test that incorrectly was setting the datetime input with seconds,
which the browser does not accept.

Change-Id: Iec8ec431185a77235f458516cce3c11e7e8eb9b6