fix(Parser): Fix LinkTarget/PageReference confusion, strengthen types
The generic `array` types resulted in array items being `mixed` and thus
Phan (and IDEs) not noticing that `Title::isSamePageAs()` was called
with a `LinkTarget` rather than a `PageReference` – which wasn’t a
problem as long as the actual value was `Title` (which implements both
interfaces), but blown up my dev wiki as soon as a `TitleValue` sneaked
in. Add stricter types to make static analysis find this, and fix the
now-highlighted error by calling `Title::isSameLinkAs()` instead.

Change-Id: I4483d410a6570af2bd2aff9dd280fd3ede687eff