Fall back to linear tests if parallel test suite generation fails
When PHPUnit tests contain errors, the test suite splitting fails
because it does not have enough information from PHPUnit to accurately
report where the failure has occurred. In this case, the only way to
get the information (with the PHPUnit 9) is to run the test suite
completely and read the failure output.

Implement a fallback in the parallel test suite generation so that
tests are run non-parallel when parallel generation fails.

Bug: T379764
Change-Id: I550b11f81c15eba01359c30e5c59db0715afd4cd