ResourceLoader: Remove unused 'thisValue' from mw.trackSubscribe
This exposed the internal queue structure, which thus means that
future changes to this structure are noticable through the public
interface (mw.trackSubscribe callbacks). Those callbacks could rely
on the presence of a specific property, or perhaps modify or merge
it with something else, and thus also rely on the absence of certain
properties that they added themselves before or after that merge.

Avoid this by removing support for it.

It was introduced in 2013 with I8c7af097e6 (873b60f194), but never
used:

https://codesearch.wmcloud.org/deployed/?q=mw.trackSubscribe%7Cthis%5C.topic%5Cb&files=js%24&excludeFiles=test%7Cnode_modules

Change-Id: I2d9bc9b7228ab5f2df2160181f41c40cce8f42b6