Clean up old Special:UserLogout GET request anti-preloading hacks
Back in the day, when Special:UserLogout would log you out on a GET
request, a workaround was implemented so that users of a particular
ISP that preloaded GET requests would not immediately become logged
out: T19790 (SVN r62967 and r86182 / e3e5cf1bb9 and 12cb599cb7).

This is no longer needed now that it requires a POST request, and
most of the code was removed in 38bbfde448, but some bits remained.
Remove the message 'suspicious-userlogout', now unused. Remove
a comment about always generating Special:UserLogout links with
an '&' character, now no longer true.

Change-Id: Ib39ef71fdfa285e30c8376527cb6c7ea177d46a4