Tests: Enforce requirement that modules loaded on page load are documented
This provides a second iteration of a recently removed test in Vector,
that we removed in I0eb22ebbed67ded008de7ca75f9f8f399a214ef8

* The new test only fails if it detects modules on page load that haven't got
associated budgets ie. are unexpected and new - this catches the issue we've
had with OOUI being accidentally added to page load.
* All budgets are defined in the associated extensions - we defer to maintainers
of those extensions to monitor and consider bundle size increases to these
modules rather than police it in Vector.
* We do not check the size of modules that are tied to wiki pages e.g.
site or site.styles - to support this bundlesizes can be defined as null to skip
testing

The downsides of this approach are:
* we are not monitoring overall size and trusting each other to be cautious
when we add new dependencies.
* we are altering the meaning of bundlesize.config.json to assume that only
modules added on page load are declared there, which is not always the case.

Bug: T360590
Depends-On: I02e096065432187d3fbb66656946ca52922ce01d
Change-Id: Ie8b524240b8eba53a24f5272e44b69a20b0ffba9