Stats: Merge preg_… calls in StatsUtils for performance
This piece of code is called exceptionally often, hundreds of
thousands of times. While it's certainly not a bottleneck it started
to stick out in the XDebug profile runs I recently did. The two regex
patterns can easily be merged and still do the same as before, as
proven by the tests.

Bug: T240685
Change-Id: I4b5ef027c2a7f0be84cd4c27760ff3f7f8f2a4ae