htmlform: Implement `max` parameter in HTMLMultiSelectField
Can be used to specify a maximum number of selectable items.
Server-side, the constraint is always checked upon form submission (in
`validate()`), regardless of other options.

Client-side, this is only implemented in the dropdown version (OOUI's
MenuTagMultiselect), because CheckboxMultiselectInputWidget does not
support any bounds on the number of selected elements.

Bug: T380406
Change-Id: I26845cc215186fc4993cd7cba84222696dc06d81