ResourceLoader: Run addOneModuleResponse() once instead of twice
In the usual request, we need to generate two outputs for the same
module response: first minified, and second unminified (for source
maps). We did that by calling addOneModuleResponse() twice with two
different minifiers. This ran the code generating the response twice.
Most of it was already cached internally, but some parts weren't,
causing log warnings like this (which led me to find this problem):

  SQLBagOStuff.WARNING: Duplicate get():
  "global:resourceloader-filter:minify-js:9:<md5>" fetched 2 times

Instead, call addOneModuleResponse() once, remember what it does,
and replay those calls on the two minifiers.

We have to remember and replay the calls, instead of doing them
immediately, because sometimes the result of the first minifier
is needed to decide whether we even construct the second minifier.

Change-Id: I2a4516ebb95973f5bb1a863f4b6eecf04edd2a92