REST: Allow response body properties to have translatable descriptions
In https://gerrit.wikimedia.org/r/c/mediawiki/core/+/1078506 we added
the ability to specify path/query parameter descriptions as
MessageValue objects to expose them in a translatable way in
OpenAPI specs. Make the same possible for response body parameters.

Caveat: only schema and top-level property descriptions are translated.
Description translation in nested objects is not supported in this patch.

Bug: T379706
Change-Id: I68215ad55f2e8b2fec4315516e874192faba6595