Revision: make RevisionRecord::getPage() return a value object
RevisionRecord::getPage() currently returns a PageIdentity, which may
be a Title or a WikiPage. In the context of an ongoing page deletion,
the state of these objects may change and the page ID may get set to 0.
To provide a reliable stable interface that can be used e.g. in a
deferred update or job, we need to return an immutable value object
instead.

NOTE: this may break callers of getPage() that downcast to Title,
or that rely on the page ID contained in the returned PageIdentity
to be updated dynamically during page creation or deletion.

Bug: T380536
Change-Id: I7440f8f30b05949842903094908232455f6a2239