SpecialBlock.vue: keep track of promises in the store, add loading state
Introduce pushPromise() and a local promises ref to the store. This
keeps track of UI-blocking API requests, ensuring the form is disabled
while the requests are awaiting response. This is meant to be
reminiscent of OOUI's pushPending(). A native JavaScript Set is used to
ensure uniqueness and without the need to name the promises.

BlockLog is reworked to use the getBlockLogData() from the store so that
we get the block log and active blocks in the same API request. The
promise is cached locally to ensure there are no duplicate requests
across BlockLog instances.

Remove ref from `details` as it is not exported or meant to be mutable.

Bug: T380437
Change-Id: I9c018f85c4ccdf5aa75a3fb24c7363e62d33f938