Add type hint to Language::getCode
The `mCode` property is set in the contructor and never changed
afterwards; it's guaranteed to be a string.

This change is not only for type safety, but also to simplify PHPUnit
tests: a return typehint allows PHPUnit to make the mocked method
return an instance of that type (instead of null) when the method is not
explicitly configured with `method()`. For a real example, see
I4d28a157aaeb4faa17f4785dccd3b591230b03d0.

Change-Id: I3293344d7e3376d9c18fd0c6372c636f5c9af881