Add $linestart to parseWithPostprocessing
The call to getText in Echo::DiscussionParser (line 1293) seems a good
fit for the usage of MessageCache::parseWithPostprocessing to replace it,
but it is called with a $linestart parameter that we can't pass to this
method anymore. Hence, we reintroduce the parameter with a default value
of true as it was used in the method, using the same argument order as
Parser::parse().

Bug: T293512
Depends-On: Iab08f6272b2995a42b0732ad82e73e6beb796cd7
Change-Id: Icdab9cec8b40baa090b450994fd82dcbb7a83afe