ParserCache: After fetching from persistent cache, save to in-process cache
In this code, $this->cache is a persistent cache (e.g. SQLBagOStuff),
and $this->metadataProcCache is an in-process cache (HashBagOStuff).
The in-process cache is supposed to be a tiny subset of the persistent
cache (see the doc comment on $metadataProcCache).

This works fine when saving new entries to both caches. But when
fetching existing entries that were not added by the current process,
it would always fetch from the persistent cache, repeatedly:

  SQLBagOStuff.WARNING: Duplicate get():
  "mediawiki:pcache:idoptions:<id>" fetched 2 times

So, on a successful fetch from the persistent cache, save the same
entry to the in-process cache.

This is almost the same thing that CachedBagOStuff provides, but it
apparently can't be used here because it caches misses (see T277829).

Bug: T277829
Change-Id: I4b2d925ec1d27223d6b135e485de53be24a099d0