Provide the page title to localization on message parsing
Message parsing expects to have a page title available. The canonical
way to do this seems to use the request context to create the message to
be parsed, but this does not seem strictly necessary:
Message::setContext only sets up the title and the user language
callback, which would end up handled in the same way if we were passing
RequestContext::getMain() here, which we have tried to avoid relying on
so far in this part of the code.
Hence, we rely on the page DBKey set by Parsoid to re-create a Title and
pass it to the Message as PageReference, so that it is available when
parsing the message.

Bug: T380045
Change-Id: I587e64bed068a33fec27a91d303fe0a8cd585317