Avoid exception on mTemplateIds/mTemplate array discrepancy
Adding ParserOutput::getLinkList surfaced cases of discrepancies between
the mTemplate and mTemplateIds fields, in particular one triggered in
Flow.
This patch aims at avoiding crashes on
getLinkList(ParserOutputLinkTypes::TEMPLATES) when that is the case by
returning 0 ("zero for broken", according to the documentation of
$mTemplateIds) in such cases.
Note that a fix in Flow (or in other places) will not fix the issue
immediately because this desynchronization has probably been cached for
a long time, and is just surfacing now.
Whether we want to log such occurrences is unclear to me - I left it out
in this patch to avoid further log spam for now.

Bug: T380862
Change-Id: Icb119c77deecfaa45f1be4bab82ed32627e6c4f4
(cherry picked from commit bbde95ad3a3f1d100749fbb7fc5483d78eb90db9)