ResourceLoader: Use MW_INSTALL_PATH constant in transformResourcePath()
This function was created in I7f38bfc1be (419db58c8b), which predates
the existence of MW_INSTALL_PATH. If created today, I would have
simply used this constant.

But, given that was created when $IP was the standard, we naturally
wrote tests for it the same way, which still predated the constant,
in Ic5a4af705c (a9d14289d7).

Then, a few years later in commit I7142af16d6 (cf581bb2ca), $IP was
replaced with two things: a constant MW_INSTALL_PATH, and a mutable
$wgBaseDirectory, on the very reasonable assumption that, if $IP
is overwritten in tests, then it could not be migrated to a constant.

At least for ResourceLoader, this turns out not to be the case. The
only reason $IP was overwritten was to shorten the path in the test
fixture. The only thing transformResourcePath() cares about is whether
a file is served within MediaWiki's install path (where it can map
the URL to a file on disk, and md5-hash that), or served externally.

In PHPUnit we want to test it both ways, and given that the PHPUnit
files are all within /w/, this is perfectly testable as-is without
needing to overwrite MW_INSTALL_PATH. It just needs a slightly longer
URL.

Change-Id: If3b7942d25f3d8380fe76f5c81faa275d2e013c2