Setup: Remove unused $wgBaseDirectory in favour of MW_INSTALL_PATH
In commit I7142af16d6 (cf581bb2ca), $IP was replaced with two things:
a constant MW_INSTALL_PATH, and a mutable $wgBaseDirectory. It offered
both on the very reasonable assumption that, given $IP is overwritten
in several places in tests, it can't be migrated to a constant.

Over the years, this turned out not the case:

1. Resources.php, switched to constant in I56c28d36e (e446a71f86).
2. GitInfo, switched to constant in Ib2c89461ed (268c2a5b5a).
3. DatabaseUpdater, switched to constant in I502bdb652e.
4. transformResourcePath(), switched to constant in If3b7942d25.

There is no other use left, and afaik it was never adopted anywhere,
other than in these 4 places in MediaWiki core as part of the initial
$IP deprecation above. As such, remove without deprecation.

Change-Id: Ia273414d77743e5c59bded793de6fdc6667fd8b8