RevisionRecord: add convenience methods for main slot
When accessing slot content and meta-data, most code wants to only
access the main slot. Add convenience methods for making this less
awkward.

Originally, the intent was for all code to support arbitrary slots. This
hasn't happened, instead we spread SlotREcord::MAIN all over the code
base. It seems better to adjust the interface of RevisionRecord to
reality.

Change-Id: I8603f95c8e39d6fc3522a47f74657798e7f7c061