user: Add newFromNameOrIp() method to UserFactory
This simplifies creation of User object for either an IP or a
username, without turning off validation for the named user,
which causes problems subsequently when other codepaths expect
named users passed to them to be valid.

Most callers that currently call User/UserFactory::newFromName()
with validation turned off can be migrated to this method, which
is safer.

Bug: T381217
Change-Id: I4a7bae534a06d279492c7e237871e3e8b4ac7167