Fix missing split group error log output
Changes introduced in I550b11f81c15eba01359c30e5c59db0715afd4cd for
T379764 resulted in the log files for the phpunit split groups
being generated with invalid names (`phpunit_output_1_1.log` instead
of `phpunit_output_1_database.log`). This meant the
PhpUnitConsoleLogOutputProcessor was no longer able to find the log
files to present a summary at the end of a failed run.

This patch fixes the bug, and also goes some way to refactoring the
code to make it more testable and debuggable.

Bug: T378481
Change-Id: I0c345ea56529e6768f80a00440e9a13aa05ad069