MWHttpRequest: allow using local proxy in cli mode
As stated in the linked task, it is not guaranteed that we can't serve
"local" requests in cli mode. This is only true if the following
assumptions are also true:
* you run the cli scripts on a different machine as the webserver
* You don't run a service mesh to mediate between MediaWiki and the
  external services

While the former is true for the wikimedia production setup, it's not a
common setup at all. On the other hand, it's been quite a few years since
Wikimedia started using a service mesh.

So it seems to me that for the absolute majority of installations, this
condtional was actually harmful.

This patch allows using the local proxy also in cli mode, and puts
control back in the hands of the system administrators.

Bug: T381251
Change-Id: I724a1e840d586db2b6b57b9990a5747290e39494