PageUpdatedEvent: explicitly model cause and performer
This introduces the concept of a "cause" and "performer" into
PageUpdatedEvent. Causes are e.g. "edit" or "import" or "move".
The performer is the same as the author for edits, but not
necessarily for undeletions, moves, etc.

Note that cause and performer were already being tracked in
DerivedPageDataUpdater for logging purposes.
PageUpdatedEvent was modeling the cause using flags.

This change formalizes the concept of "cause" and "performer"
and ensures it is handled consistently.

As an aside, this simplifies some test cases in
DerivedPageDataUpdaterTest to avoid confused state.

Bug: T383031
Bug: T383095
Change-Id: I79bde5604790636c91848f265fffd81d6348a8b4