installer: Avoid empty handler for LoadExtensionSchemaUpdates
Passing a falsy value for the handler is hard-deprecated. Apparently,
the deprecation notice is hard to miss, because it's been there for
months.

Also fix installer code where some methods were assuming an array to be
a map of hook names to handlers, and others were assuming the same
variable to be a list of handlers for the LoadExtensionSchemaUpdates
hook (bug introduced in I2c41ece04aed11a176559d4364cea2f3320b9574).

Bug: T358446
Change-Id: I18bae539c8572e1186a68492a257c0c50eeaff65