objectcache: Automatically fallback to the second db in line
...when the first attempt causes db error.

We currently have this problem that if a parsercache host goes down,
until it's replaced or fixed, all 1/nth pageviews trigger a reparse
(unless cached in memcached).

With a rather simple change, if db error is returned, simply fallback to
the second database in line, that way in case a pc host goes down, we
will have a graceful degradation and move the load to the rest of hosts
automatically. It's similar to ring designs.

This helps us re-architecture parsercache in favor of getting rid of
idle spares because we wouldn't need an immediate fix in case of issues.

It also can be used to make x2 much more robust (we can set up x4 and
they would automatically fail over in case of issues).

Bug: T373037
Change-Id: I95424f453396fbcdd399e588a4b6638ecb572441